+++ addon/mixins/reload-list-mixin.js
@@ -43,6 +43,7 @@ export default Ember.Mixin.create({
    * @param {String} [options.page] Current page.
    * @param {String} [options.sorting] Current sorting.
    * @param {String} [options.filter] Current filter.
+   * @param {String} [options.filterProjectionName] Name of model projection which should be used for filtering throught search-element on toolbar. Filtering is processed only by properties defined in this projection.
    * @param {String} [options.predicate] Predicate to limit records.
    * @return {Promise}  A promise, which is resolved with a set of loaded records once the server returns.
    */
@@ -73,6 +74,15 @@ export default Ember.Mixin.create({
     let projection = Ember.get(modelConstructor, 'projections')[projectionName];
     if (!projection) {
       throw new Error(`No projection with '${projectionName}' name defined in '${modelName}' model.`);
+    }
+
+    let filterProjectionName = reloadOptions.filterProjectionName;
+    let filterProjection = undefined;
+    if (filterProjectionName) {
+      filterProjection = Ember.get(modelConstructor, 'projections')[filterProjectionName];
+      if (!filterProjection) {
+        throw new Error(`No projection with '${filterProjection}' name defined in '${modelName}' model.`);
+      }
     }
 
     let allPredicates = Ember.A();
@@ -131,7 +141,11 @@ export default Ember.Mixin.create({
 
     const filter = reloadOptions.filter;
     const filterCondition = reloadOptions.filterCondition;
-    const filterPredicate = filter ? this._getFilterPredicate(projection, { filter, filterCondition }) : undefined;
+    const filterPredicate = filter 
+            ? this._getFilterPredicate(
+                      filterProjection ? filterProjection : projection,
+                      { filter, filterCondition }) 
+            : undefined;
     allPredicates.addObject(filterPredicate);
     allPredicates = allPredicates.compact();
     const resultPredicate = allPredicates.length > 1 ?